refactor: change LoginUser use case file to maintain naming pattern
