+++ src/modules/user/useCases/login/__tests__/LoginUserController.spec.ts
@@ -2,7 +2,7 @@ import { mock } from 'jest-mock-extended'
 
 import { badRequest, ok, serverError } from '../../../../../shared/infra/http/http'
 import { Validator } from '../../../../../validation/Validator'
-import { LoginUser } from '../LoginUse'
+import { LoginUser } from '../LoginUser'
 import { LoginUserController, LoginUserControllerRequestDTO } from '../LoginUserController'
 import { LoginUserErrors } from '../LoginUserErrors'
