+++ src/modules/user/useCases/register/RegisterUserController.spec.ts
@@ -4,7 +4,7 @@ import { mock } from 'jest-mock-extended'
 import userFactory from '../../../../../tests/factories/userFactory'
 import { badRequest, conflict, created } from '../../../../shared/infra/http/http'
 import { Validator } from '../../../../validation/Validator'
-import { LoginUser } from '../login/LoginUse'
+import { LoginUser } from '../login/LoginUser'
 import { RegisterUser } from './RegisterUser'
 import { RegisterUserController, RegisterUserControllerRequestDTO } from './RegisterUserController'
 import { RegisterUserErrors } from './RegisterUserErrors'
